+++ app/context/AudioContext.tsx
@@ -3,7 +3,7 @@ import React, { createContext, useContext, useMemo, useRef, useState } from 'rea
 import { Verse } from '@/types';
 import { RECITERS, Reciter } from '@/lib/reciters';
 
-// This is from the new CleanPlayer component.
+// This is from the new AudioPlayer component.
 export type RepeatOptions = {
   mode: 'off' | 'single' | 'range' | 'surah';
   start?: number;
@@ -92,7 +92,17 @@ export const AudioProvider = ({ children }: { children: React.ReactNode }) => {
       isPlayerVisible,
       closePlayer,
     }),
-    [playingId, isPlaying, loadingId, activeVerse, repeatOptions, reciter, volume, playbackRate, isPlayerVisible]
+    [
+      playingId,
+      isPlaying,
+      loadingId,
+      activeVerse,
+      repeatOptions,
+      reciter,
+      volume,
+      playbackRate,
+      isPlayerVisible,
+    ]
   );
 
   return <AudioContext.Provider value={value}>{children}</AudioContext.Provider>;
