+++ app/dev/player/page.tsx
@@ -1,28 +1,32 @@
 'use client';
 
 import React, { useState } from 'react';
-import { CleanPlayer } from '@/app/components/player';
-import type { Track, RepeatOptions } from '@/app/components/player/types';
+import { AudioPlayer } from '@/app/components/player';
+import type { Track, RepeatOptions } from '@/app/components/player/AudioPlayer';
 
 const DEMO_TRACKS: Track[] = [
   {
     id: 'alafasy-001',
-    url: 'https://download.quranicaudio.com/quran/mishaari_raashid_al_afasy/001.mp3',
+    src: 'https://download.quranicaudio.com/quran/mishaari_raashid_al_afasy/001.mp3',
     title: 'Al-Fātiḥah - Mishary Alafasy',
-    reciter: { id: 1, name: 'Mishary Alafasy', path: 'mishaari_raashid_al_afasy' },
+    artist: 'Mishary Alafasy',
+    coverUrl: '',
+    durationSec: 0,
   },
   {
     id: 'minshawi-001',
-    url: 'https://download.quranicaudio.com/quran/muhammad_siddeeq_al-minshaawee/mujawwad/001.mp3',
+    src: 'https://download.quranicaudio.com/quran/muhammad_siddeeq_al-minshaawee/mujawwad/001.mp3',
     title: 'Al-Fātiḥah - Minshawi',
-    reciter: { id: 2, name: 'Minshawi (Mujawwad)', path: 'muhammad_siddeeq_al-minshaawee' },
+    artist: 'Minshawi (Mujawwad)',
+    coverUrl: '',
+    durationSec: 0,
   },
 ];
 
 export default function Page() {
   const [track, setTrack] = useState<Track>(DEMO_TRACKS[0]);
   const [repeat, setRepeat] = useState<RepeatOptions>({
-    mode: 'none',
+    mode: 'off',
     start: 0,
     end: 0,
     repeatEach: 1,
@@ -49,7 +53,7 @@ export default function Page() {
           >
             {DEMO_TRACKS.map((t) => (
               <option key={t.id} value={t.id}>
-                {t.reciter?.name}
+                {t.artist}
               </option>
             ))}
           </select>
@@ -66,7 +70,7 @@ export default function Page() {
             }
             className="border p-1 rounded"
           >
-            <option value="none">None</option>
+            <option value="off">Off</option>
             <option value="single">Single</option>
             <option value="range">Range</option>
           </select>
@@ -114,10 +118,10 @@ export default function Page() {
           </label>
         </div>
         <pre className="bg-gray-100 p-2 rounded text-xs">
-          {JSON.stringify({ reciter: track.reciter, repeat }, null, 2)}
+          {JSON.stringify({ track, repeat }, null, 2)}
         </pre>
       </div>
-      <CleanPlayer src={track.url} title={track.title} />
+      <AudioPlayer track={track} repeatOptions={repeat} onRepeatChange={setRepeat} />
     </div>
   );
 }
